refactor: FollowButtonをファイルに切り出す
- UserDetailActivityの内部にあったFollowButton関数をファイルに切り出した
- setFollowState関数をリネームした
